requirements.txt
gitpython
transformers
typing_extensions
requests_cache
tokenizers
torch
huggingface_hub
datasets